Moves TestRestClient to client package
