+++ spring-cloud-gateway-server-mvc/src/test/java/org/springframework/cloud/gateway/server/mvc/test/client/JsonPathAssertions.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.cloud.gateway.server.mvc.test;
+package org.springframework.cloud.gateway.server.mvc.test.client;
 
 import java.util.function.Consumer;
